+++ src/App.vue
@@ -1,82 +1,5 @@
 <template lang="pug">
-div(class="general")
-    div(class="aside")
-      div(class="Proj")
-        img(src="./assets/Logo@3x.svg" width="32" height="30" alt="logo" class="logo")
-        div
-          h2
-            | PROJECTUS
-      div(class="avaka")
-        img(:src='User.avatar' width="48" height="48" alt="Avatar" class="Avatar")
-        ul
-          li(class="name")
-            | {{User.name}}
-          li(class="status")
-            | {{User.tag}}
-        div(class="dotdotdot_avka")
-          div(class="dot")
-          div(class="dot")
-          div(class="dot")
-      div(class="compl_task")
-        ul(class="Completed" @click="TaskChanger")
-          li(class="number")
-            | {{compl}}
-          li(class="Tasks")
-            | Completed Tasks
-        ul(class="Open")
-          li(class="number")
-            | {{open}}
-          li(class="Tasks")
-            | Open Tasks
-      div(class="menu")
-        h2
-          | MENU
-        div(class="Notif")
-          p
-        ul
-          li
-            | Home
-          li
-            | My Tasks
-          li
-            | Notifications
-    div(class="gside")
-      div(class="header")
-        div(class="title")
-          img(class="Shapes" src="./assets/Shapes@2x.png" alt="logo" height="40" width="40")
-          h1
-            | Website Redesign
-          div(class="dotdotdot_title")
-            div(class="dot_position")
-              div(class="dot")
-              div(class="dot")
-              div(class="dot")
-          div(class="space")
-          div(class="users")
-            img(src="./assets/Avatar_s.svg" height="30" width="30" alt="avatar")
-            img(src="./assets/Avatar_s.svg" height="30" width="30" alt="avatar")
-            img(src="./assets/Avatar_s.svg" height="30" width="30" alt="avatar")
-          div(class="buttons")
-            div(class="Share")
-              p
-                | Share
-            div(class="Chat")
-              div
-                | Chat
-        div(class="chapters")
-          nav(class="nav")
-            router-link(class="n_button" to = '/tasks')
-              | Tasks
-            router-link(class="n_button" to = '/kanban')
-              | Kanban
-            router-link(class="n_button" to = '/activity')
-              | Activity
-            router-link(class="n_button" to = '/calendar')
-              | Calendar
-            router-link(class="n_button" to = '/files')
-              | Files
-      div(class="content_zone")
-        router-view
+router-view
 </template>
 
 <script lang="ts">
@@ -85,26 +8,7 @@ export default defineComponent(
   {
     name: 'App',
     components:{},
-    data (): Record<string, any>{
-      return {
-        User: {
-          avatar: require('./assets/Avatar.svg'),
-          name: 'Jean Gonzales',
-          tag: 'Product Owner',
-        },
-        compl: 328,
-        open: 10,
-      }
-    },
-    methods: {
-      TaskChanger(){
-        const sumbit = confirm('Are you sure you want to change the number of tasks?')
-        if (sumbit && this.open > 0){
-          this.compl++
-          this.open--
-        }
-      }
-    }
+
   }
 )
 </script>
