+++ src/types/Task_interface.ts
@@ -1,0 +1,3 @@
+export interface Task_interface {
+  name: string
+}
